Remove unnecessary changes
