+++ api/server/handlers/user/github_callback.go
@@ -85,6 +85,7 @@ func (p *UserOAuthGithubCallbackHandler) ServeHTTP(w http.ResponseWriter, r *htt
 	// non-fatal send email verification
 	if !user.EmailVerified {
 		err = startEmailVerification(p.Config(), w, r, user)
+
 		if err != nil {
 			p.HandleAPIErrorNoWrite(w, r, apierrors.NewErrInternal(err))
 		}
@@ -147,11 +148,13 @@ func upsertUserFromToken(config *config.Config, tok *oauth2.Token) (*models.User
 			}
 
 			user, err = config.Repo.User().CreateUser(user)
+
 			if err != nil {
 				return nil, err
 			}
 
 			err = addUserToDefaultProject(config, user)
+
 			if err != nil {
 				return nil, err
 			}
