+++ api/server/handlers/user/google_callback.go
@@ -88,6 +88,7 @@ func (p *UserOAuthGoogleCallbackHandler) ServeHTTP(w http.ResponseWriter, r *htt
 	// non-fatal send email verification
 	if !user.EmailVerified {
 		err = startEmailVerification(p.Config(), w, r, user)
+
 		if err != nil {
 			p.HandleAPIErrorNoWrite(w, r, apierrors.NewErrInternal(err))
 		}
@@ -133,11 +134,13 @@ func upsertGoogleUserFromToken(config *config.Config, tok *oauth2.Token) (*model
 			}
 
 			user, err = config.Repo.User().CreateUser(user)
+
 			if err != nil {
 				return nil, err
 			}
 
 			err = addUserToDefaultProject(config, user)
+
 			if err != nil {
 				return nil, err
 			}
