+++ internal/billing/metronome.go
@@ -20,16 +20,6 @@ const (
 	defaultCollectionMethod = "charge_automatically"
 	defaultMaxRetries       = 10
 	porterStandardTrialDays = 15
-
-	// referralRewardRequirement is the number of referred users required to
-	// be granted a credits reward
-	referralRewardRequirement = 5
-	// defaultRewardAmountCents is the default amount in USD cents rewarded to users
-	// who reach the reward requirement
-	defaultRewardAmountCents = 1000
-	// defaultPaidAmountCents is the amount paid by the user to get the credits
-	// grant, if set to 0 it means they were free
-	defaultPaidAmountCents = 0
 )
 
 // MetronomeClient is the client used to call the Metronome API
